+++ packages/lib/src/components/Card/services/ClickToPayService.ts
@@ -2,6 +2,7 @@ import { ISrcInitiator } from './sdks/AbstractSrcInitiator';
 import { CheckoutResponse, IsRecognizedResponse } from './types';
 import { ISrcSdkLoader } from './sdks/SrcSdkLoader';
 import { SecureRemoteCommerceInitResult } from './configMock';
+import { createShopperCardsList } from './utils';
 
 export enum CtpState {
     Idle = 'Idle',
@@ -166,36 +167,19 @@ class ClickToPayService implements IClickToPayService {
     }
 
     private async getSecureRemoteCommerceProfile(idTokens: string[]): Promise<void> {
-        function createShopperMaskedCardsData(memo, srcProfile) {
-            const { profiles, srcCorrelationId } = srcProfile;
-
-            const cards = profiles.reduce((memo, profile) => {
-                const profileCards = profile.maskedCards.map(maskedCard => ({
-                    dateOfCardLastUsed: maskedCard.dateOfCardLastUsed,
-                    panLastFour: maskedCard.panLastFour,
-                    srcDigitalCardId: maskedCard.srcDigitalCardId,
-                    paymentCardDescriptor: maskedCard.paymentCardDescriptor,
-                    srcCorrelationId
-                }));
-                return [...memo, ...profileCards];
-            }, []);
-
-            return [...memo, ...cards];
-        }
-
         const srcProfilesPromises = this.sdks.map(sdk => sdk.getSrcProfile(idTokens));
         const srcProfiles = await Promise.all(srcProfilesPromises);
 
         console.log('srcProfiles', srcProfiles);
 
-        const data = srcProfiles.reduce(createShopperMaskedCardsData, []);
-        console.log('cards', data);
-        this.shopperCards = data;
+        const cards = createShopperCardsList(srcProfiles);
+        console.log('cards', cards);
+        this.shopperCards = cards;
 
         // TODO: verify when APi return multiple profiles. What to do with that?
-        // For now it is taking only the first one of the first response
-        this.srcProfile = srcProfiles[0]?.profiles[0];
-        this.srcCorrelationId = srcProfiles[0]?.srcCorrelationId;
+        // // For now it is taking only the first one of the first response
+        // this.srcProfile = srcProfiles[0]?.profiles[0];
+        // this.srcCorrelationId = srcProfiles[0]?.srcCorrelationId;
 
         this.setState(CtpState.Ready);
     }
